requirements.txt
openai
chromadb
beautifulsoup4
requests
nltk
fastapi
uvicorn
python-dotenv
